Move fixtures and constants from conftest to modules
